% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_go.R
\name{barplot_go}
\alias{barplot_go}
\title{Bar plot for gene ontology visualization.}
\usage{
barplot_go(
  go_data,
  type = "cluego",
  top = "50",
  go_process = "",
  min_genes = "",
  header = ""
)
}
\arguments{
\item{go_data}{Input GO data to visualize.}

\item{type}{Default cluego. Parameter to specify input data source.}

\item{top}{Default value is 50. Parameter to set top number of processes to be visualized.}

\item{go_process}{Turned off by default. Parameter to add a specific keyword to only visuzalize GO terms that contains it.}

\item{min_genes}{Turned off by default. Parameter to set threshold for biological processes containing minimum number of genes.}

\item{header}{Set a title for the barplot.}
}
\value{
A bar plot of gene ontology results.
}
\description{
This function visualizes gene ontology results from ClueGO as bar plot.
}
