% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_luc.R
\name{barplot_luc}
\alias{barplot_luc}
\title{This function visualizes qPCR results as bar plot.}
\usage{
barplot_luc(
  luc_data,
  type = "glomax",
  group1 = "",
  group2 = "",
  group3 = "",
  group4 = "",
  group5 = "",
  stat = TRUE,
  test = "t-test",
  generate_table = FALSE
)
}
\arguments{
\item{luc_data}{Input qPCR data to visualize.}

\item{type}{Default biorad. Specify model of thermocycler.}

\item{group1}{Control group.}

\item{group2}{Target group.}

\item{group3}{Target group.}

\item{group4}{Target group.}

\item{group5}{Target group.}

\item{stat}{Default TRUE. Enable or disable statistical analysis.}

\item{test}{Default t-test. Select statistical testing method.}

\item{generate_table}{Default False. Set to True to create a table with the results everytime run the function.}

\item{ref1}{Reference gene one.}

\item{ref2}{Reference gene two.}

\item{goi}{Gene of interest.}

\item{tech_rep}{Number of technical replicates. Optional if consideration of technical replicates for statistical testing is desired.}
}
\value{
A bar plot of qPCR results.
}
\description{
This function visualizes qPCR results as bar plot.
}
