% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbleplot_GO.R
\name{bubbleplot_go}
\alias{bubbleplot_go}
\title{Bubble plot for gene ontology visualization.}
\usage{
bubbleplot_go(go_data, top = " ", go_process = " ")
}
\arguments{
\item{go_data}{Input GO data to visualize.}

\item{top}{Turned off by default. Parameter to set top amount of processes to be visualized.}

\item{go_process}{Turned off by default. Parameter to add a specific keyword to only visuzalize GO terms that contains it.}
}
\value{
A bubble plot of gene ontology results.
}
\description{
This function visualizes gene ontology results from ClueGO as bubble plot.
}
