% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maplot_dge.R
\docType{data}
\name{dge_data}
\alias{dge_data}
\title{MA-plot visualization of DGE analysis.}
\format{
An object of class \code{character} of length 1.
}
\usage{
dge_data
}
\arguments{
\item{dge_data}{Path to the input file.}

\item{FDR}{Default 0.05. Adjust FDR value.}

\item{FC}{Default 2. Adjust log2FC threshold.}

\item{TOP}{Default 10. Adjust top number of DE genes to visualize.}

\item{header}{Default is empty. Set a title for the MA-plot.}

\item{type}{Default deseq2. Specify input datatype.}

\item{mapping}{Default is FALSE. Set to TRUE to convert the gene identifiers to HGNC.}
}
\value{
An ma-plot of RNA-Seq DGE data.
}
\description{
This function visualizes DGE results as an MA-plot.
}
\keyword{datasets}
