% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot_ipms.R
\name{dotplot_ipms}
\alias{dotplot_ipms}
\title{Dot plot for IP-MS experiment visualization.}
\usage{
dotplot_ipms(ip_data)
}
\arguments{
\item{ip_data}{Path to the input file}
}
\value{
A dot plot of the ip_data
}
\description{
This function visualizes IP-MS results with 2 axis
for number of validated peptides in 2 samples.
}
