% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_dge.R
\name{heatmap_dge}
\alias{heatmap_dge}
\title{Heatmap for differential gene expression data.}
\usage{
heatmap_dge(heatmap_data, List = " ")
}
\arguments{
\item{heatmap_data}{Path to the input file.}

\item{List}{Path to the list of genes file which will be visualized.}
}
\value{
A heatmap of heatmap_data.
}
\description{
This function visualizes differential gene expresion data
as heatmap.
}
