% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_hgnc.R
\docType{data}
\name{map_data}
\alias{map_data}
\title{Auto mapping to HGNC of a variety of gene identifiers.}
\format{
An object of class \code{character} of length 1.
}
\usage{
map_data
}
\arguments{
\item{map_data}{Path to the input file.}
}
\value{
A table containing HGNC gene IDs.
}
\description{
This function automatically detects a variety of gene identifiers and converts them to HGNC using biomaRt.
}
\keyword{datasets}
