% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maplot_RNASeq.R
\name{maplot_rnaseq}
\alias{maplot_rnaseq}
\title{MA-plot visualization of RNA-Seq DGE analysis.}
\usage{
maplot_rnaseq(rnaseq_data, FDR = 0.01, FC = 0.5)
}
\arguments{
\item{rnaseq_data}{Path to the input file.}

\item{FDR}{Default 0.01. Adjust FDR value.}

\item{FC}{Default 0.5. Adjust log2FC threshold.}
}
\value{
An ma-plot of RNA-Seq DGE data.
}
\description{
This function visualizes RNA-Seq DGE results as an MA-plot.
}
