% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ipms.R
\name{plot_ipms}
\alias{plot_ipms}
\title{Dot plot for IP-MS experiment visualization.}
\usage{
plot_ipms(ip_data)
}
\arguments{
\item{ip_data}{Path to the input file}
}
\value{
A dot plot of the  ip_data
}
\description{
This function visualizes IP-MS results with 2 axis
for number of validated peptides in NANOS-IP vs pCMV6-IP
when i run the script manually it generates proper plot
but when i run it as a function, it proceeds without error
and creates pdf file but the file doesnt open.
}
