% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violinplot.R
\name{violinplot}
\alias{violinplot}
\title{Violin plot to visualize biological data.}
\usage{
violinplot(violin_data, x_lab = "conditions", y_lab = "values")
}
\arguments{
\item{violin_data}{Path to the input file.}

\item{x_lab}{Default is "conditions". Parameter to set label for conditions.}

\item{y_lab}{Default is "values". Parameter to set label for values.}
}
\value{
A violin plot of the input data.
}
\description{
This function visualize biological data as a violin plot
}
