% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_go.R
\name{barplot_go}
\alias{barplot_go}
\title{Bar plot for gene ontology visualization.}
\usage{
barplot_go(go_data, top = " ", go_process = " ")
}
\arguments{
\item{go_data}{Input GO data to visualize.}

\item{top}{Turned off by default. Parameter to set top amount of processes to be visualized.}

\item{go_process}{Turned off by default. Parameter to add a specific keyword to only visuzalize GO terms that contains it.}
}
\value{
A bar plot of gene ontology results.
}
\description{
This function visualizes gene ontology results from ClueGO as bar plot.
}
