% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_qpcr.R
\name{barplot_qpcr}
\alias{barplot_qpcr}
\title{Analyze and visualize qPCR results with barplot.}
\usage{
barplot_qpcr(
  qpcr_data,
  type = "biorad",
  ref = "",
  target = "",
  stat = "t-test"
)
}
\arguments{
\item{qpcr_data}{Input qPCR data to visualize.}

\item{top}{Turned off by default. Parameter to set top amount of processes to be visualized.}

\item{go_process}{Turned off by default. Parameter to add a specific keyword to only visuzalize GO terms that contains it.}
}
\value{
A bar plot of gene ontology results.
}
\description{
This function visualizes qPCR results as bar plot.
}
