% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_qpcr.R
\name{barplot_qpcr}
\alias{barplot_qpcr}
\title{Analyze and visualize qPCR results with barplot.}
\usage{
barplot_qpcr(
  qpcr_data,
  type = "biorad",
  ref = "",
  target = "",
  stat = "t-test"
)
}
\arguments{
\item{qpcr_data}{Input qPCR data to visualize.}

\item{type}{}

\item{ref}{}

\item{stat}{Default t-test.}
}
\value{
A bar plot of qPCR results.
}
\description{
This function visualizes qPCR results as bar plot.
}
