% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_qpcr.R
\name{barplot_qpcr}
\alias{barplot_qpcr}
\title{This function visualizes qPCR results as bar plot.}
\usage{
barplot_qpcr(
  qpcr_data,
  type = "biorad",
  group1 = "",
  group2 = "",
  ref1 = "",
  ref2 = "",
  goi = "",
  tech_rep = 4,
  stat = "t-test"
)
}
\arguments{
\item{qpcr_data}{Input qPCR data to visualize.}

\item{type}{Default Bio-rad. Specify model of thermocycler.}

\item{group1}{Control group.}

\item{group2}{Target group.}

\item{ref1}{Reference gene one.}

\item{ref2}{Reference gene two.}

\item{goi}{Gene of interest.}

\item{tech_rep}{Number of technical replicates. Necessary for accurate statistical calculation}

\item{stat}{Default t-test. Select statistical testing method.}
}
\value{
A bar plot of qPCR results.
}
\description{
This function visualizes qPCR results as bar plot.
}
